Refactor Retirements Overview cards
